Address review feedback changes
